Bugfixes, small interface changes and testing
I switched some attributes of the helper classes to not be "public"/immediately accessible. Also fixed a big that was introduced in .to_dense() due to the change in _get_block/_set_block.
I added a bunch of tests that should cover the regular use-cases. Everything passing on CPU & GPU.