Number of bugfixes concerning element-wise access in nnz distribution and some GPU issues.
I added a sectioning strategy parameter to the get_section_sizes function. This returns the correct section sizes for stack and nnz in our convention. Also noticed that the dtranspose on GPU did actually not work correctly before.

Finally adjusted the tests and tweaked the tested sparsity pattern a bit to get a more awkward nnz than always 120.