Refactor rgf_dist and add obc_blocks
I assimilated the RGFDist to the old way of doing RGF. There now is a _serinv backend module that has all the Schur-based methods and the reduced system. I also added the OBCBlocks things to use that correctly there.

I still need to propagate the return_retarded keyword correctly and update the docstrings